enables page aware navbar that highlights when on the page
The previous solution where I added a new class specific to a page
causes problem because of the way we enforce responsive navbar. As
we are targetting specific divs (#navContainer and #topNavbar) and
modify their classes manually with a toggler function, we can't
just append new classes to those two specific ids without also
modifying our toggler function.

It is not a problem, but it introduces complexity, as now we have
to edit in multiple places to tell our stylesheet that we're in a
certain page. By adding a new class to our div container, and then
modify our function that targets it to enable responsiveness, and
then ensuring that they're all the correct classes they're adding,
would be prone of errors.

Archlinux did it beautifully. Why can't we?

Turned out we can. It evades me for quite some time. However I am
glad that I am still resourceful enough to find a solution. What I
did is basically add a new wrapper container for navbar. This time,
I use a nav element to wrap our navbar, and give it a class
specific to a certain page or page group. Then our responsive
styling works out of the box, and our navbar is context aware.

How beautiful.